requirements.txt
# Core ML Libraries
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# HuggingFace Transformers
transformers>=4.35.0
accelerate>=0.24.0
huggingface-hub>=0.16.0

# Image Processing
opencv-python>=4.8.0
Pillow>=9.5.0

# Utilities
numpy>=1.21.0
protobuf>=4.0.0

# Optional: For better performance
# safetensors>=0.3.0